requirements.txt
aiogram==3.21.0
pydantic==2.8.2
anyio==4.9.0
APScheduler==3.11.0
colorama==0.4.6
numpy==1.25.2
pandas==2.0.3
pause==0.3
python-dotenv==1.1.1
pytz==2024.1
requests==2.32.4
rich==14.1.0
selenium==4.34.2

ta==0.11.0
tabulate==0.9.0
Telethon==1.40.0
toml==0.10.2
tzlocal==5.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
